Travail sur la suppression des contraintes. Première version trop laxiste: 7 modules ne sont plus sur le lieux d'origine, il faudrait restreindre les prochains choix au lieu sélectionné pour les modules dont la contrainte n'a pas été respectée. Essai avec un propagate en cours : lors de la boucle de recherche le propagate permet de prendre une décision lorsqu'on est en cours de recherche Pour l'instant le propagate ne fonctionne pas dans Choco.
